refactor(DejaTextMetricsModule): Move text-metrics to src/common/core
