+++ src/common/core/text-metrics/index.ts
