+++ src/common/core/text-metrics/readme.md
