+++ src/component/tree-list/index.ts
@@ -10,11 +10,11 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { MatButtonModule, MatCheckboxModule, MatInputModule } from '@angular/material';
+import { DejaTextMetricsModule } from '../../common/core/text-metrics/index';
 import { DejaItemModule } from './../../common/core/item-list/index';
 import { DejaChildValidatorModule } from './../../common/core/validation/index';
 import { DejaDragDropModule } from './../dragdrop/index';
 import { DejaListLoaderModule } from './../loaders/index';
-import { DejaTextMetricsModule } from './text-metrics/index';
 import { DejaTreeListComponent } from './tree-list.component';
 
 @NgModule({
@@ -41,4 +41,3 @@ export class DejaTreeListModule { }
 
 export * from './tree-list-scroll-event';
 export * from './tree-list.component';
-export * from './text-metrics/index';
